refactor(hub): move partition ops + vault-diff → with-cargo, repoint /cargo
git mv the partition-transfer ops (extract/adopt/decrypt-partition,
walk-closure, describe-extraction) and vault-diff.ts into src/with-cargo/,
alongside the #537 cargo barrel (moved from src/cargo/). The /cargo export
surface is unchanged (golden frozen) — only the barrel's vault-diff import
becomes local. extract-partition.ts moves verbatim (import paths only).
Pure file moves plus import-path rewrites; no behavior change.
