+++ packages/hub/__tests__/carry-ledger.test.ts
@@ -16,9 +16,9 @@ import { hashEntry } from '../src/with-commit/history/ledger/entry.js'
 import { envelopePayloadHash } from '../src/with-commit/history/ledger/hash.js'
 import type { LedgerEntry } from '../src/with-commit/history/ledger/entry.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
-import { reKeyLedger, reKeyClosure, extractPartition } from '../src/with-share/bundle/extract-partition.js'
-import { adoptPartition, createOwnerOnAdoptedPartition } from '../src/with-share/bundle/adopt-partition.js'
-import { readNoydbBundle, parseExtractedPartitionBody } from '../src/with-share/bundle/bundle.js'
+import { reKeyLedger, reKeyClosure, extractPartition } from '../src/with-cargo/extract-partition.js'
+import { adoptPartition, createOwnerOnAdoptedPartition } from '../src/with-cargo/adopt-partition.js'
+import { readNoydbBundle, parseExtractedPartitionBody } from '../src/with-pod/bundle.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
