+++ packages/hub/__tests__/custodian-role.test.ts
@@ -16,7 +16,7 @@ import { createNoydb } from '../src/noydb.js'
 import type { Noydb } from '../src/noydb.js'
 import { checkGate, PolicyDeniedError } from '../src/policy/index.js'
 import { putCredential } from '../src/with-party/team/sync-credentials.js'
-import { extractPartition } from '../src/with-share/bundle/extract-partition.js'
+import { extractPartition } from '../src/with-cargo/extract-partition.js'
 
 function inlineMemory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
