+++ packages/hub/__tests__/extract-partition.test.ts
@@ -15,8 +15,8 @@ import { ref } from '../src/refs.js'
 import { ConflictError } from '../src/errors.js'
 import { decrypt, base64ToBuffer, generateDEK } from '../src/crypto.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
-import { reKeyClosure, sealDeks, extractPartition } from '../src/with-share/bundle/extract-partition.js'
-import { readNoydbBundle, readNoydbBundleHeader, parseExtractedPartitionBody } from '../src/with-share/bundle/bundle.js'
+import { reKeyClosure, sealDeks, extractPartition } from '../src/with-cargo/extract-partition.js'
+import { readNoydbBundle, readNoydbBundleHeader, parseExtractedPartitionBody } from '../src/with-pod/bundle.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
