+++ packages/hub/__tests__/managed-mode-adoption.test.ts
@@ -11,8 +11,8 @@ import { ConflictError } from '../src/errors.js'
 import { MemorySealingKeyProvider } from '../src/with-party/team/managed-passphrase.js'
 import { shamirRecoveryProvider } from '@noy-db/on-shamir'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
-import { extractPartition } from '../src/with-share/bundle/extract-partition.js'
-import { adoptPartition, createOwnerOnAdoptedPartition } from '../src/with-share/bundle/adopt-partition.js'
+import { extractPartition } from '../src/with-cargo/extract-partition.js'
+import { adoptPartition, createOwnerOnAdoptedPartition } from '../src/with-cargo/adopt-partition.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
