+++ packages/hub/__tests__/partition-handed-over-ledger.test.ts
@@ -8,7 +8,7 @@ import { createNoydb } from '../src/noydb.js'
 import { withHistory } from '../src/with-commit/history/index.js'
 import { ConflictError } from '../src/errors.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
-import { extractPartition } from '../src/with-share/bundle/extract-partition.js'
+import { extractPartition } from '../src/with-cargo/extract-partition.js'
 
 function memory(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
